Create math module with tests
